requirements.txt
opencv_python==4.5.5.64
albumentations==0.5.1 --no-binary=imgaug,albumentations
ray>=1.0.1
einops==0.3.0
kornia==0.7.3
loguru==0.5.3
yacs>=0.1.8
tqdm
autopep8
pylint
ipython
jupyterlab
matplotlib
h5py
pytorch-lightning==1.5.10
joblib>=1.0.1
Pillow==9.5.0
pymagsac==0.2.0
scikit-learn
